Allow position updates to be delayed
Having a unique delay like this helps us prevent deadlocks. We
were getting an issue with large feature imports where each new
feature created was kicking of a whole new position update query
so we were updating potentially thousands of features constantly.
This was causing deadlocks in a lot of cases and making it so
all features were not imported for no obvious reason (to the user).

This is really the only place I could find to put this. The problem
is that each feature creation when importing is it's own API call
so we need some state outside of the server to keep track of the
uniqueness of the operation. Leveraging delayed resque for this
is the most straight forward approach.